Do not force default `default_user` when falling back to non-HTTP auth
If the Authorization HTTP header is missing, we want to fallback to
the credentials in the CONNECT frame. Therefore, we must not force the
default `default_user` credentials in the `#stomp_configuration{}`
record.

This problem went unnoticed before and the authentication always failed
because the default `default_user` was made of Erlang strings instead
of binaries, which was rejected by the Erlang client and RabbitMQ. This
problem was fixed in commit a49e814e70e2ee839d282b9614cf8bfbab87acb8 in
rabbitmq-comp.

While here, remove trailing whitespaces.